[code-infra] Create custom rule for undefined in optional props
This rule naively (currently) checks for all optional properties in
interface and types and provides autofix to add `| undefined` wherever
it doesn't exist.
One slight optimization is added where locally defined types are also
checked for `| undefined` and if it already has `undefined`, its not
added at places where the type is used.